+++ src/tailwind/feedback/index.ts
@@ -1,2 +1,3 @@
 export * from './CardModal';
 export * from './ModalDialog';
+export * from './Result';